Fix behavior with header X-Forward-HTTPS: off
In this case the internal scheme in Apache was
set unconditionally (the mere presence of the
header is treated as if it's value was always 'on').

This affects e.g. the DirectorySlash Redirect in
mod_dir (the one adding trailing slashes)
and leads to redirects like this:

http://host.name/foo -> https://host.name:80/foo/

when the initial request contained a header

X-Forwarded-HTTPS: off

or

X-HTTPS: off