Handle X-Request-Id header or generate request id
If a trusted reverse proxy is making the request and it contains a
header (name configurable) X-Request-Id the id is extracted from the
header and stored in the subprocess environment as X_REQUEST_ID.

Otherwise a request id is generated based on the connection id and the
timestamp in microseconds.

In any case (if this feature is turned on) the X-Request-Id header is
set with the trustworthy request id.